feat: replace quarkus-world
