+++ quarkus-world/.dockerignore
@@ -2,3 +2,4 @@
 !target/*-runner
 !target/*-runner.jar
 !target/lib/*
+!target/quarkus-app/*
