+++ quarkus-world/.gitignore
@@ -1,3 +1,10 @@
+#Maven
+target/
+pom.xml.tag
+pom.xml.releaseBackup
+pom.xml.versionsBackup
+release.properties
+
 # Eclipse
 .project
 .classpath
@@ -28,9 +35,5 @@ nb-configuration.xml
 *.orig
 *.rej
 
-# Maven
-target/
-pom.xml.tag
-pom.xml.releaseBackup
-pom.xml.versionsBackup
-release.properties
+# Local environment
+.env
