+++ quarkus-world/src/main/docker/Dockerfile.native
@@ -1,21 +1,24 @@
 ####
-# This Dockerfile is used in order to build a container that runs the Quarkus application in native (no JVM) mode
+# This Dockerfile is used in order to build a container that runs the Quarkus application in native (no JVM) mode.
 #
-# Before building the docker image run:
+# Before building the container image run:
 #
-# mvn package -Pnative -Dquarkus.native.container-build=true
+# ./mvnw package -Pnative
 #
 # Then, build the image with:
 #
-# docker build -f src/main/docker/Dockerfile.native -t quarkus/quarkus-world .
+# docker build -f src/main/docker/Dockerfile.native -t quarkus/quarkus-ws-world .
 #
 # Then run the container using:
 #
-# docker run -i --rm -p 8080:8080 quarkus/quarkus-world
+# docker run -i --rm -p 8080:8080 quarkus/quarkus-ws-world
 #
 ###
-FROM registry.access.redhat.com/ubi8/ubi-minimal:8.1
+FROM registry.access.redhat.com/ubi8/ubi-minimal:8.5
 WORKDIR /work/
+RUN chown 1001 /work \
+    && chmod "g+rwX" /work \
+    && chown 1001:root /work
 COPY --chown=1001:root target/*-runner /work/application
 
 EXPOSE 8080
