+++ quarkus-world/src/main/java/org/acme/getting/started/AppStartup.java
@@ -1,4 +1,4 @@
-package org.acme;
+package org.acme.getting.started;
 
 import javax.websocket.ContainerProvider;
 import javax.enterprise.context.ApplicationScoped;
