+++ quarkus-world/src/main/java/org/acme/getting/started/Config.java
@@ -1,4 +1,6 @@
-package org.acme.getting.started;///////////////////////////////////////////////////////////////////////////////
+package org.acme.getting.started;
+
+///////////////////////////////////////////////////////////////////////////////
 // Config.java Fall 2018
 //
 // This file contains constants used within your Eliza program.
