+++ quarkus-world/src/main/resources/application.properties
@@ -1,5 +1,5 @@
-# Configuration file
-# key = value
 quarkus.log.level=INFO
-
 quarkus.native.resources.includes=Quarkus.rsp
+
+chat.hub.username=quarkus
+chat.hub.host=ws://localhost:11800/chat/
