--- quarkus-world/src/test/java/org/acme/getting/started/NativeMessagingResourceIT.java
@@ -1,9 +1,0 @@
-package org.acme.getting.started;
-
-import io.quarkus.test.junit.NativeImageTest;
-
-@NativeImageTest
-public class NativeMessagingResourceIT extends MessagingResourceTest {
-
-    // Execute the same tests but in native mode.
-}
