--- quarkus-ws-world/src/main/resources/application.properties
@@ -1,5 +1,0 @@
-quarkus.log.level=INFO
-quarkus.native.resources.includes=Quarkus.rsp
-
-chat.hub.username=quarkus
-chat.hub.host=ws://localhost:11800/chat/
